// client/src/app/_models/passport.ts
export interface Passport {
    access_token: string,
    display_name: string,
    avatar_url?: string
    xp?: number
    level?: number
}

export interface RegisterModel {
    username: string
    password: string
    display_name: string
}
export interface LoginModel {
    username: string
    password: string
}